Show tx fees, restructure TransactionSummary
