+++ src/components/TransactionSummary/TransactionSummary.tsx
@@ -1,24 +1,20 @@
 import React from "react"
-import { Memo, Operation, Transaction, Horizon } from "stellar-sdk"
+import { Operation, Transaction } from "stellar-sdk"
 import Divider from "@material-ui/core/Divider"
-import MuiListItem from "@material-ui/core/ListItem"
-import MuiListItemIcon from "@material-ui/core/ListItemIcon"
-import MuiListItemText from "@material-ui/core/ListItemText"
-import Tooltip from "@material-ui/core/Tooltip"
-import CheckIcon from "@material-ui/icons/Check"
-import UpdateIcon from "@material-ui/icons/Update"
-import WarningIcon from "@material-ui/icons/Warning"
 import { useAccountDataSet } from "../../hooks"
 import { Account, AccountsContext } from "../../context/accounts"
 import { SignatureRequest } from "../../lib/multisig-service"
-import { getAllSources, getSignerKey, signatureMatchesPublicKey } from "../../lib/stellar"
+import { getAllSources } from "../../lib/stellar"
 import { isPotentiallyDangerousTransaction } from "../../lib/transaction"
-import { ObservedAccountData } from "../../subscriptions"
-import { warningColor } from "../../theme"
-import { HorizontalLayout } from "../Layout/Box"
-import { List, ListItem } from "../List"
-import PublicKey from "../PublicKey"
+import { List } from "../List"
 import OperationListItem from "./Operations"
+import {
+  DangerousTransactionWarning,
+  Signers,
+  SourceAccount,
+  TransactionMemo,
+  TransactionMetadata
+} from "./Transaction"
 
 function makeOperationSourceExplicit(
   operation: Operation,
@@ -31,114 +27,6 @@ function makeOperationSourceExplicit(
   return effectiveSource === transaction.source && (effectiveSource === localAccountPubKey || !localAccountPubKey)
     ? operation
     : { ...operation, source: effectiveSource }
-}
-
-function MetaDetails(props: { children: React.ReactNode }) {
-  return <div style={{ fontSize: "80%", marginTop: 8, marginLeft: 16 }}>{props.children}</div>
-}
-
-function TransactionMemo(props: { memo: Memo; style?: React.CSSProperties }) {
-  if (props.memo.type === "none" || !props.memo.value) return null
-
-  const memo = typeof props.memo.value === "string" ? props.memo.value : props.memo.value.toString("hex")
-  const typeLabel = props.memo.type.substr(0, 1).toUpperCase() + props.memo.type.substr(1)
-
-  return <ListItem heading={`${typeLabel} Memo`} primaryText={<MetaDetails>{memo}</MetaDetails>} style={props.style} />
-}
-
-function SignerStatus(props: { hasSigned: boolean; style?: React.CSSProperties }) {
-  const Icon = props.hasSigned ? CheckIcon : UpdateIcon
-  return (
-    <Tooltip title={props.hasSigned ? "Has signed the transaction" : "Awaiting their signature"}>
-      <Icon style={{ opacity: props.hasSigned ? 1 : 0.5, ...props.style }} />
-    </Tooltip>
-  )
-}
-
-// tslint:disable-next-line no-shadowed-variable
-const Signer = React.memo(function Signer(props: {
-  hasSigned: boolean
-  signer: Horizon.AccountSigner | { key: string; weight: number }
-  transaction: Transaction
-}) {
-  return (
-    <HorizontalLayout alignItems="center">
-      <>
-        <SignerStatus hasSigned={props.hasSigned} style={{ marginRight: 8 }} />
-        <div style={{ whiteSpace: "nowrap" }}>
-          <PublicKey
-            publicKey={getSignerKey(props.signer)}
-            style={{ display: "inline-block", fontWeight: "normal", minWidth: 480 }}
-            variant="full"
-          />
-        </div>
-      </>
-    </HorizontalLayout>
-  )
-})
-
-function Signers(props: {
-  accounts: Account[]
-  accountData: ObservedAccountData
-  transaction: Transaction
-  style?: React.CSSProperties
-}) {
-  // TODO: We should not get the signers from the source account data, but either
-  //       a) from the signature request or
-  //       b) by taking the tx source and all operation source accounts into account
-  const headingDetails = props.accountData.signers.every(signer => signer.weight === 1)
-    ? `${props.accountData.thresholds.high_threshold} of ${props.accountData.signers.length} multi-signature`
-    : `Custom consensus multi-signature`
-  return (
-    <ListItem
-      heading={`Signers (${headingDetails})`}
-      primaryText={
-        <MetaDetails>
-          <>
-            {props.accountData.signers.map(signer => (
-              <Signer
-                key={getSignerKey(signer)}
-                hasSigned={props.transaction.signatures.some(signature =>
-                  signatureMatchesPublicKey(signature, getSignerKey(signer))
-                )}
-                signer={signer}
-                transaction={props.transaction}
-              />
-            ))}
-          </>
-        </MetaDetails>
-      }
-      style={props.style}
-    />
-  )
-}
-
-function SourceAccount(props: { transaction: Transaction; style?: React.CSSProperties }) {
-  return (
-    <ListItem
-      heading="Source Account"
-      primaryText={
-        <MetaDetails>
-          <PublicKey publicKey={props.transaction.source} style={{ fontWeight: "normal" }} variant="full" />
-        </MetaDetails>
-      }
-      style={props.style}
-    />
-  )
-}
-
-function DangerousTransactionWarning(props: { style?: React.CSSProperties }) {
-  return (
-    <MuiListItem style={{ background: warningColor, ...props.style }}>
-      <MuiListItemIcon>
-        <WarningIcon />
-      </MuiListItemIcon>
-      <MuiListItemText
-        primary="Transaction initiated by unrecognized account"
-        secondary="Please review carefully. In case of doubt, prefer dismissing."
-      />
-    </MuiListItem>
-  )
 }
 
 interface TransactionSummaryProps {
@@ -208,6 +96,7 @@ function TransactionSummary(props: TransactionSummaryProps) {
         />
       ) : null}
       {props.showSource ? <SourceAccount transaction={props.transaction} style={noHPaddingStyle} /> : null}
+      <TransactionMetadata style={noHPaddingStyle} transaction={props.transaction} />
     </List>
   )
 }
